:recycle: refactor : 채팅 목록 리팩토링
