--- src/pages/friends/friendsHeader.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const FriendsHeader = () => {
-  return <FriendsHeadContainer></FriendsHeadContainer>;
-};
-
-export default FriendsHeader;
-
-const FriendsHeadContainer = styled.div``;
